move completion to common package
