+++ golem-cli/src/completion.rs
