+++ golem-cli/src/oss.rs
@@ -14,7 +14,6 @@
 
 pub mod clients;
 pub mod command;
-pub mod completion;
 pub mod factory;
 pub mod main;
 pub mod model;
