+++ golem-cli/src/oss/command.rs
@@ -19,8 +19,8 @@ use crate::command::profile::ProfileSubCommand;
 use crate::command::worker::{OssWorkerUriArg, WorkerSubcommand};
 use crate::diagnose;
 use crate::model::{ComponentUriArg, Format, HasFormatConfig, HasVerbosity};
-use crate::oss::completion;
-use crate::oss::completion::PrintCompletion;
+use crate::completion;
+use crate::completion::PrintCompletion;
 use crate::oss::model::OssContext;
 use clap::{CommandFactory, Parser, Subcommand};
 use clap_complete::Shell;
