+++ golem-cli/src/oss/main.rs
@@ -19,7 +19,7 @@ use crate::factory::ServiceFactory;
 use crate::init::{init_profile, DummyProfileAuth, ProfileAuth};
 use crate::model::{ApiDefinitionId, ApiDefinitionVersion, GolemError, GolemResult};
 use crate::oss::command::{GolemOssCommand, OssCommand};
-use crate::oss::completion::PrintCompletion;
+use crate::completion::PrintCompletion;
 use crate::oss::factory::OssServiceFactory;
 use crate::oss::model::OssContext;
 use crate::stubgen::handle_stubgen;
